change: renamed folder :art:
should be lib
