+++ src/lib/products.json
